docs: create roadmap of future features to add
While developing the project, there will be certain areas of code which are plainly obvious to be
optimized, or abstracted away, to create more flexible and modular code. However, doing this
optimizations prematurely might backfire. Therefore, try to first create an MVP and then come back
to remove roadmapped items.